change column type to numeric

+++ database.sql
@@ -24,7 +24,7 @@ CREATE TABLE IF NOT EXISTS "user_settings"
   "reinvest" boolean DEFAULT false,
   "reinvest_ratio" integer DEFAULT 0,
   "max_trade" boolean DEFAULT false,
-  "max_trade_size" integer DEFAULT 0,
+  "max_trade_size" numeric(32,8) DEFAULT 0,
   "profit_reset" timestamp
 );
